+++ src/mympd_api/jukebox.h
@@ -13,5 +13,5 @@
 void mympd_api_jukebox_clear(struct t_list *list, sds partition_name);
 bool mympd_api_jukebox_rm_entries(struct t_list *list, struct t_list *positions, sds partition_name, sds *error);
 sds mympd_api_jukebox_list(struct t_partition_state *partition_state, struct t_stickerdb_state *stickerdb,
-        sds buffer, enum mympd_cmd_ids cmd_id, long request_id, long offset, long limit, sds expression, const struct t_tags *tagcols);
+        sds buffer, enum mympd_cmd_ids cmd_id, unsigned request_id, unsigned offset, unsigned limit, sds expression, const struct t_tags *tagcols);
 #endif
