+++ src/mympd_api/last_played.h
@@ -9,7 +9,7 @@
 
 #include "src/lib/mympd_state.h"
 
-bool mympd_api_last_played_add_song(struct t_partition_state *partition_state, long last_played_count, int song_id);
+bool mympd_api_last_played_add_song(struct t_partition_state *partition_state, unsigned last_played_count, unsigned song_id);
 sds mympd_api_last_played_list(struct t_partition_state *partition_state, struct t_stickerdb_state *stickerdb,
-        sds buffer, long request_id, long offset, long limit, sds expression, const struct t_tags *tagcols);
+        sds buffer, unsigned request_id, unsigned offset, unsigned limit, sds expression, const struct t_tags *tagcols);
 #endif
