+++ src/mympd_api/lyrics.h
@@ -9,5 +9,5 @@
 
 #include "src/lib/mympd_state.h"
 
-sds mympd_api_lyrics_get(struct t_lyrics *lyrics, sds music_directory, sds buffer, long request_id, sds uri);
+sds mympd_api_lyrics_get(struct t_lyrics *lyrics, sds music_directory, sds buffer, unsigned request_id, sds uri);
 #endif
