+++ src/mympd_api/mounts.h
@@ -9,7 +9,7 @@
 
 #include "src/lib/mympd_state.h"
 
-sds mympd_api_mounts_list(struct t_partition_state *partition_state, sds buffer, long request_id);
-sds mympd_api_mounts_neighbor_list(struct t_partition_state *partition_state, sds buffer, long request_id);
-sds mympd_api_mounts_urlhandler_list(struct t_partition_state *partition_state, sds buffer, long request_id);
+sds mympd_api_mounts_list(struct t_partition_state *partition_state, sds buffer, unsigned request_id);
+sds mympd_api_mounts_neighbor_list(struct t_partition_state *partition_state, sds buffer, unsigned request_id);
+sds mympd_api_mounts_urlhandler_list(struct t_partition_state *partition_state, sds buffer, unsigned request_id);
 #endif
