+++ src/mympd_api/song.h
@@ -10,6 +10,6 @@
 #include "src/lib/mympd_state.h"
 
 sds mympd_api_song_details(struct t_mympd_state *mympd_state, struct t_partition_state *partition_state,
-    sds buffer, long request_id, const char *uri);
-sds mympd_api_song_comments(struct t_partition_state *partition_state, sds buffer, long request_id, const char *uri);
+    sds buffer, unsigned request_id, const char *uri);
+sds mympd_api_song_comments(struct t_partition_state *partition_state, sds buffer, unsigned request_id, const char *uri);
 #endif
