+++ src/mympd_api/webradios.h
@@ -17,8 +17,8 @@ bool mympd_api_webradio_save(sds workdir, sds name, sds uri, sds uri_old,
         sds genre, sds picture, sds homepage, sds country, sds language,
         sds codec, int bitrate, sds description, sds state);
 bool mympd_api_webradio_delete(sds workdir, struct t_list *filenames);
-sds mympd_api_webradio_get(sds workdir, sds buffer, long request_id, sds filename);
-sds mympd_api_webradio_list(sds workdir, sds buffer, long request_id, sds searchstr,
-        long offset, long limit);
+sds mympd_api_webradio_get(sds workdir, sds buffer, unsigned request_id, sds filename);
+sds mympd_api_webradio_list(sds workdir, sds buffer, unsigned request_id, sds searchstr,
+        unsigned offset, unsigned limit);
 
 #endif
